Backend/requirements.txt
Flask
Flask-CORS
pymongo
python-dotenv
spotipy
deepface
opencv-python
numpy
matplotlib
tf-keras
google-generativeai